feat: refactor ConfigureToolModal and update references in ToolGrid and ToolList components
